# [complex] torch.sqrt: fix edge values