Fixes #47358 

Replace the optimized path with a slower but correct `map(std::sqrt)`

Benchmark posted below in comments.

cc: @dylanbespalko (original author of fast-path)