@kshitij12345 2020-11-05
Benchmark

Before the change :
```
[-------- sqrt torch.complex64 --------]
                       |  torch.sqrt CPU
8 threads: -----------------------------
      (32,)            |        3.0     
      (32, 32)         |        4.3     
      (2, 16, 32)      |        4.3     
      (2, 16, 32, 32)  |       13.2     

Times are in microseconds (us).
 
Using pytorch 1.8.0a0+f19637e
10 / 10[------- sqrt torch.complex128 --------]
                       |  torch.sqrt CPU
8 threads: -----------------------------
      (32,)            |        3.0     
      (32, 32)         |        7.1     
      (2, 16, 32)      |        7.1     
      (2, 16, 32, 32)  |       24.4     

Times are in microseconds (us).
```

After the change:
```
[-------- sqrt torch.complex64 --------]
                       |  torch.sqrt CPU
8 threads: -----------------------------
      (32,)            |        3.7     
      (32, 32)         |       37.8     
      (2, 16, 32)      |       38.4     
      (2, 16, 32, 32)  |      160.6     

Times are in microseconds (us).

[------- sqrt torch.complex128 --------]
                       |  torch.sqrt CPU
8 threads: -----------------------------
      (32,)            |        3.5     
      (32, 32)         |       26.9     
      (2, 16, 32)      |       26.7     
      (2, 16, 32, 32)  |      115.9     

Times are in microseconds (us).
```

Benchmark Script
```python
import torch
import itertools
import time
from torch.utils.benchmark import Timer
from torch.utils.benchmark import Compare
import sys

print('Using pytorch %s' % (torch.__version__))

shapes = [(32,), (32, 32), (2, 16, 32), (2, 16, 32, 32)]
results = []
num_threads = 8
dtype = torch.cdouble
repeats = 10

for mat1_shape in shapes:
    mat1 = torch.rand(*mat1_shape, dtype=dtype, device='cpu')

    tasks = [("torch.sqrt(mat1)", "torch.sqrt CPU")]
    timers = [Timer(stmt=stmt, num_threads=num_threads, label=f"sqrt {dtype}", sub_label=f"{(mat1_shape)}", description=label, globals=globals()) for stmt, label in tasks]

    for i, timer in enumerate(timers * repeats):
        results.append(
            timer.blocked_autorange()
        )
        print(f"\r{i + 1} / {len(timers) * repeats}", end="")
        sys.stdout.flush()
comparison = Compare(results)
comparison.print()
```