@kshitij12345 2020-11-05
> https://github.com/pytorch/pytorch/issues/47358#issuecomment-722252021
> Issue looks to arise from use of `abs`,
> 
> ```python
> import numpy as np
> import torch
> 
> print(torch.__version__)
> x = torch.tensor([0.-1.0000e+20j] * 10)
> print("ABS:", x.abs())
> print("ABS:", np.abs(x.numpy()))
> ```
> 
> Output
> 
> ```
> 1.8.0a0+f19637e
> ABS: tensor([       inf,        inf,        inf,        inf,        inf,        inf,
>                inf,        inf, 1.0000e+20, 1.0000e+20])
> ABS: [1.e+20 1.e+20 1.e+20 1.e+20 1.e+20 1.e+20 1.e+20 1.e+20 1.e+20 1.e+20]
> ```



> Issue looks to arise from use of `abs`,
> 
> ```python
> import numpy as np
> import torch
> 
> print(torch.__version__)
> x = torch.tensor([0.-1.0000e+20j] * 10)
> print("ABS:", x.abs())
> print("ABS:", np.abs(x.numpy()))
> ```
> 
> Output
> 
> ```
> 1.8.0a0+f19637e
> ABS: tensor([       inf,        inf,        inf,        inf,        inf,        inf,
>                inf,        inf, 1.0000e+20, 1.0000e+20])
> ABS: [1.e+20 1.e+20 1.e+20 1.e+20 1.e+20 1.e+20 1.e+20 1.e+20 1.e+20 1.e+20]
> ```