@samestep 2020-11-05
It looks like your PR broke a few other tests (besides the performance regressions you documented); these are the ones I'm seeing from the [pytorch_linux_xenial_py3_6_gcc5_4_test](https://app.circleci.com/pipelines/github/pytorch/pytorch/235495/workflows/ffd1d7bc-87a3-407f-912f-81988cc6a9f8/jobs/8718910) job:
```
Nov 05 08:21:01 [  FAILED  ] 4 tests, listed below:
Nov 05 08:21:01 [  FAILED  ] SqrtAndReciprocal/2.Sqrt, where TypeParam = at::vec256::(anonymous namespace)::Vec256<c10::complex<float> >
Nov 05 08:21:01 [  FAILED  ] SqrtAndReciprocal/3.Sqrt, where TypeParam = at::vec256::(anonymous namespace)::Vec256<c10::complex<double> >
Nov 05 08:21:01 [  FAILED  ] InverseTrigonometric/2.Asin, where TypeParam = at::vec256::(anonymous namespace)::Vec256<c10::complex<float> >
Nov 05 08:21:01 [  FAILED  ] InverseTrigonometric/2.ACos, where TypeParam = at::vec256::(anonymous namespace)::Vec256<c10::complex<float> >
Nov 05 08:21:01 
Nov 05 08:21:01  4 FAILED TESTS
```
Could you fix all the broken tests to give better context of the scope of your proposed change?