@mruberry 2020-11-06
The failure on MacOS looks real, the failure on Windows seems unrelated.

The test should go in test_unary_ufuncs.py and have a comment linking the issue it's testing for. 

You can file a new issue for the fail on MacOS, skip the test on MacOS, and add a comment linking to THAT issue to explain the skip. 

The performance regression is regrettable but it's better to be correct than fast and wrong.