@kshitij12345 2020-11-10
>  The test should go in test_unary_ufuncs.py and have a comment linking the issue it's testing for.

Done

> You can file a new issue for the fail on MacOS, skip the test on MacOS, and add a comment linking to THAT issue to explain > the skip.

Done