@mruberry 2020-11-11
Now that this has been merged we should review speeding it up again in a way that doesn't have numeric issues.