@kshitij12345 2020-11-11
> Now that this has been merged we should review speeding it up again in a way that doesn't have numeric issues.

Yes.
Probably doing something based on the following comment might help.

> Did a bit of digging. seems like this has nothing to do with the vec256 but more of the fact that std::sqrt was implemented as (aa + bb)^0.5 where as the it should've been using something similar to hypot, e.g. |a| * (1+(b/a)^2)^0.5 .

Reference to comment : https://github.com/pytorch/pytorch/issues/47500#issuecomment-723504390